fix: app crashing on selecting date, optimized viewModel and used data… (#2835)

+++ core/model/src/commonMain/kotlin/org/mifos/mobile/core/model/entity/templates/loans/Product.kt
@@ -117,7 +117,7 @@ data class Product(
 
     val maxTrancheCount: Int? = null,
 
-    val principalThresholdForLastInstallment: Int? = null,
+    val principalThresholdForLastInstallment: Double? = null,
 
     val holdGuaranteeFunds: Boolean? = null,
